Fix 3 test bugs: framework double-count, optimizer miscompile, integer overflow UB
- TEST macro unconditionally incremented pass count even on failure;
  now only counts pass when fail count is unchanged
- counted_realloc inlined at -O1+ let the optimizer eliminate the NULL
  check on realloc(ptr, SIZE_MAX); mark noinline to prevent this
- test_ncache_hash multiplied int by 31415 causing signed overflow UB;
  use unsigned arithmetic for well-defined wrapping

+++ test/test_unit_devhash.h
@@ -11,7 +11,8 @@
 
 /* ncache hash from rnmh.c */
 static int test_ncache_hash(unsigned long inode, unsigned long addr, int mod) {
-    return ((((int)(addr >> 2) + (int)(inode)) * 31415) & (mod - 1));
+    return ((int)(((unsigned int)((unsigned int)(addr >> 2) + (unsigned int)(inode)) * 31415u) &
+                  (unsigned int)(mod - 1)));
 }
 
 TEST(sfhash_range) {
